Catch correct exception in MailChimp client
This wouldn't have caught the exception class changing when the
dependency was updated (we'd need something like VCR for that) but this
code was untested and now it's not!
